# storage/openapi/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bugloos API",
        "description": "Bugloos API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/v1/user": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get list of Users",
                "operationId": "user_get",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bugloos"
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "example": "admin"
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "count": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "724b650c-da5e-4301-886c-3a508a68f3a1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Administrator"
                    },
                    "username": {
                        "type": "string",
                        "example": "admin"
                    },
                    "register": {
                        "type": "string",
                        "example": "1400/08/03 | 15:20"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "auth": {
                "type": "http",
                "description": "Use auth/login to get the JWT token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "jwt",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}